#1 gameid van int naar long veranderd in ListEntry en ListEntity

+++ app/src/main/java/com/example/nexus/data/db/ListEntity.kt
@@ -7,7 +7,7 @@ import com.example.nexus.ui.routes.list.ListCategory
 @Entity(tableName = "list")
 data class ListEntity (
     @PrimaryKey(autoGenerate = true) val id: Int,
-    val gameId: Int,
+    val gameId: Long,
     val title: String,
     val score: Int,
     val hoursPlayed: Int,
